Move Gradle test layout to separated project
