+++ gradle/libraries/gradle-test-script/build.gradle
@@ -1,0 +1,12 @@
+plugins {
+	id 'java-library'
+}
+
+java {
+	sourceCompatibility = JavaVersion.VERSION_1_8
+	targetCompatibility = JavaVersion.VERSION_1_8
+}
+
+dependencies {
+	implementation 'dev.gradleplugins:gradle-build-script'
+}
