+++ gradle/libraries/gradle-test-script/src/main/java/dev/gradleplugins/testscript/TestIncludedBuild.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.nvm.fixtures;
+package dev.gradleplugins.testscript;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
@@ -21,23 +21,31 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.function.Consumer;
 
-public final class TestBuildSrc extends AbstractTestGradleBuild<TestBuildSrc> implements TestGradleBuild {
+public final class TestIncludedBuild extends AbstractTestGradleBuild<TestIncludedBuild> implements TestGradleBuild {
 	private final TestGradleBuild parent;
+	private final String buildPath;
 
-	public TestBuildSrc(TestGradleBuild parent, Path location) {
+	private TestIncludedBuild(TestGradleBuild parent, Path location, String buildPath) {
 		super(location);
 		this.parent = parent;
+		this.buildPath = buildPath;
 	}
 
-	static TestBuildSrc newInstance(TestGradleBuild parent) {
+	static TestIncludedBuild newInstance(TestGradleBuild parent, String buildPath) {
 		try {
-			final Path location = parent.getLocation().resolve("buildSrc");
-			Files.createDirectories(location);
-			// Note: buildSrc doesn't require a settings.gradle[.kts]
-			return new TestBuildSrc(parent, location);
+			final Path location = parent.getLocation().resolve(buildPath);
+			Files.createDirectories(location); // ensure all directories exists
+
+			// Note: included builds require a settings.gradle[.kts] even if empty
+			Files.createFile(location.resolve("settings.gradle"));
+			return new TestIncludedBuild(parent, location, buildPath);
 		} catch (IOException e) {
 			throw new UncheckedIOException(e);
 		}
+	}
+
+	public String buildPath() {
+		return buildPath;
 	}
 
 	public void parentBuild(Consumer<? super TestGradleBuild> action) {
