+++ gradle/libraries/gradle-test-script/src/main/java/dev/gradleplugins/testscript/TestLayout.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.nvm.fixtures;
+package dev.gradleplugins.testscript;
 
 import java.nio.file.Path;
