+++ subprojects/nokee-version-management/src/functionalTest/java/dev/nokee/nvm/NokeeArtifactRepositoryFunctionalTest.java
@@ -16,11 +16,9 @@
 package dev.nokee.nvm;
 
 import dev.gradleplugins.buildscript.blocks.RepositoriesBlock;
-import dev.gradleplugins.buildscript.statements.Statement;
-import dev.gradleplugins.buildscript.syntax.Syntax;
 import dev.gradleplugins.runnerkit.GradleRunner;
 import dev.nokee.internal.testing.junit.jupiter.ContextualGradleRunnerParameterResolver;
-import dev.nokee.nvm.fixtures.TestLayout;
+import dev.gradleplugins.testscript.TestLayout;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectory;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectoryExtension;
 import org.junit.jupiter.api.Test;
