+++ subprojects/nokee-version-management/src/functionalTest/java/dev/nokee/nvm/NokeeVersionManagementServiceUsesVersionFromSystemPropertyFunctionalTest.java
@@ -17,7 +17,7 @@ package dev.nokee.nvm;
 
 import dev.gradleplugins.runnerkit.GradleRunner;
 import dev.nokee.internal.testing.junit.jupiter.ContextualGradleRunnerParameterResolver;
-import dev.nokee.nvm.fixtures.TestLayout;
+import dev.gradleplugins.testscript.TestLayout;
 import lombok.val;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectory;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectoryExtension;
