+++ subprojects/nokee-version-management/src/functionalTest/java/dev/nokee/nvm/NokeeVersionManagementVersionPatternSmokeTest.java
@@ -18,7 +18,7 @@ package dev.nokee.nvm;
 import dev.gradleplugins.runnerkit.GradleRunner;
 import dev.nokee.internal.testing.junit.jupiter.ContextualGradleRunnerParameterResolver;
 import dev.nokee.internal.testing.junit.jupiter.GradleAtLeast;
-import dev.nokee.nvm.fixtures.TestLayout;
+import dev.gradleplugins.testscript.TestLayout;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectory;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectoryExtension;
 import org.junit.jupiter.api.Test;
